payment-images/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
aiofiles==0.7.0
jinja2==3.0.1
motor==2.5.1
python-magic==0.4.27